+++ taylor.cpp
@@ -8,7 +8,7 @@ class TaylorGenerator : public Halide::Generator<TaylorGenerator> {
 public:
     Input<int> nsamples {"nsamples"};
     Input<float> S_L {"S_L"};
-    Output<Buffer<float>> output{"output", 1};
+    Output<Buffer<double>> output{"output", 1};
 
     void generate() {
         output = taylor_func(nsamples, S_L);
